+++ developer/agent/app/main.9.0.0.9728fac1702f0f4a.js
@@ -1,0 +1,1 @@
+import"./agent.14487cb5.js";import"mithril";import"localforage";