+++ developer/agent/app/main.9.0.0.aed4566a78089bcb.js
@@ -1,1 +1,0 @@
-import"./agent.dd811507.js";import"mithril";import"localforage";