+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ## [7.10.0] - 2023-05-22
 ### Added
 - 新增支援 importmap 
+- (template) webgame4 webgame6 使用 importmap
 
 ### Changed
 - 更新 agent 版本